Adding timezone and missing field error handling code.
Note: GH timestamps are always UTC, so it is safe to delete
the tz info. Commits are local time, so we first convert to
UTC then delete the info to keep it compatible with the previously or separately crawled data.

For None-s in string fields, late and early handling has been
added. Early, we try to replace at an update. Late, we replace when recreating the tf-idf model.